[DP-324] refactor: providerwrapper 채팅 초기화 로직 별도 컴포넌트로 분리
